cleanup plugin tests
